messed up select and input
