+++ src/components/UI/Input/Input.jsx
@@ -3,17 +3,14 @@ import { styled } from '@mui/material/styles'
 import { forwardRef } from 'react'
 
 const Input = forwardRef(
-   (
-      { value, onChange, placeholder, invalid, error, type = 'text', ...rest },
-      ref
-   ) => {
+   ({ value, onChange, placeholder, error, type = 'text', ...rest }, ref) => {
       return (
          <InputStyled
             value={value}
-            ref={ref}
             onChange={onChange}
             placeholder={placeholder}
             error={error}
+            ref={ref}
             type={type}
             {...rest}
          />
@@ -24,26 +21,28 @@ export default Input
 
 const InputStyled = styled(TextField)(() => ({
    '.MuiOutlinedInput-root': {
-      height: '3rem',
-      width: '26rem',
+      height: '42px',
+      width: '414px',
       fontFamily: 'Manrope',
       borderRadius: '8px',
+
       '& fieldset': {
          borderColor: '#D9D9D9',
       },
+
       '&:hover fieldset': {
          borderColor: '#959595',
       },
+
       '&:active fieldset': {
          borderColor: '#rgba(4, 135, 65, 0.80)',
       },
-      '&:invalid fieldset': {
-         border: '1px solid #F91515',
-      },
+
       '&.css-9ddj71-MuiInputBase-root-MuiOutlinedInput-root.Mui-disabled .MuiOutlinedInput-notchedOutline ':
          {
             border: '1px solid #F91515',
          },
+
       '&.Mui-focused fieldset': {
          borderColor: 'rgba(4, 135, 65, 0.80)',
       },
